Packaged everything into one package named botty
