+++ src/main/java/botty/commands/Command.java
@@ -1,0 +1,13 @@
+package botty.commands;
+
+import botty.exceptions.BottyException;
+
+import botty.tasks.TaskManager;
+
+public abstract class Command {
+    public abstract String execute(TaskManager taskManager, ParsedInput parsedInput) throws BottyException;
+
+    public boolean isExit() {
+        return false;
+    }
+}
