+++ src/main/java/botty/exceptions/ArgumentNotFoundException.java
@@ -1,0 +1,7 @@
+package botty.exceptions;
+
+public class ArgumentNotFoundException extends BottyException {
+    public ArgumentNotFoundException(String argument) {
+        super("The following argument is not found: " + argument);
+    }
+}
