+++ src/main/java/botty/exceptions/BottyException.java
@@ -1,0 +1,7 @@
+package botty.exceptions;
+
+public class BottyException extends Exception {
+    public BottyException(String message) {
+        super(message);
+    }
+}
