+++ src/main/java/botty/exceptions/CorruptedTaskStringException.java
@@ -1,0 +1,7 @@
+package botty.exceptions;
+
+public class CorruptedTaskStringException extends BottyException {
+    public CorruptedTaskStringException() {
+        super("Task string is corrupted! Failed to retrieve task");
+    }
+}
