+++ src/main/java/botty/exceptions/EmptyArgumentException.java
@@ -1,0 +1,7 @@
+package botty.exceptions;
+
+public class EmptyArgumentException extends BottyException {
+    public EmptyArgumentException(String flag) {
+        super("The following argument is empty: " + flag);
+    }
+}
