+++ src/main/java/botty/exceptions/EmptyCommandException.java
@@ -1,0 +1,7 @@
+package botty.exceptions;
+
+public class EmptyCommandException extends BottyException {
+    public EmptyCommandException() {
+        super("commands.Command cannot be empty!");
+    }
+}
