+++ src/main/java/botty/exceptions/IncorrectDateFormatException.java
@@ -1,0 +1,7 @@
+package botty.exceptions;
+
+public class IncorrectDateFormatException extends BottyException {
+    public IncorrectDateFormatException(String message) {
+        super(message);
+    }
+}
