+++ src/main/java/botty/exceptions/TaskListEmptyException.java
@@ -1,0 +1,7 @@
+package botty.exceptions;
+
+public class TaskListEmptyException extends BottyException {
+    public TaskListEmptyException() {
+        super("Your list is empty! Add a task with the todo, deadline or event command.");
+    }
+}
