+++ src/main/java/botty/exceptions/UnknownCommandException.java
@@ -1,0 +1,7 @@
+package botty.exceptions;
+
+public class UnknownCommandException extends BottyException {
+    public UnknownCommandException(String command) {
+        super("I'm sorry, that is not a command I am familiar with. (" + command + ")");
+    }
+}
