+++ src/main/java/botty/ui/UI.java
@@ -1,0 +1,47 @@
+package botty.ui;
+
+import java.util.Scanner;
+
+public class UI {
+    private static final String LOGO = " ________  ________  _________  _________    ___    ___                 \n" +
+            "|\\   __  \\|\\   __  \\|\\___   ___\\\\___   ___\\ |\\  \\  /  /|                \n" +
+            "\\ \\  \\|\\ /\\ \\  \\|\\  \\|___ \\  \\_\\|___ \\  \\_| \\ \\  \\/  / /                \n" +
+            " \\ \\   __  \\ \\  \\\\\\  \\   \\ \\  \\     \\ \\  \\   \\ \\    / /                 \n" +
+            "  \\ \\  \\|\\  \\ \\  \\\\\\  \\   \\ \\  \\     \\ \\  \\   \\/  /  /                  \n" +
+            "   \\ \\_______\\ \\_______\\   \\ \\__\\     \\ \\__\\__/  / /                    \n" +
+            "    \\|_______|\\|_______|    \\|__|      \\|__|\\___/ /                     \n" +
+            "                                           \\|___|/                      \n" +
+            "                                                                        \n" +
+            " _________  ___  ___  _______           ________  ________  _________   \n" +
+            "|\\___   ___\\\\  \\|\\  \\|\\  ___ \\         |\\   __  \\|\\   __  \\|\\___   ___\\ \n" +
+            "\\|___ \\  \\_\\ \\  \\\\\\  \\ \\   __/|        \\ \\  \\|\\ /\\ \\  \\|\\  \\|___ \\  \\_| \n" +
+            "     \\ \\  \\ \\ \\   __  \\ \\  \\_|/__       \\ \\   __  \\ \\  \\\\\\  \\   \\ \\  \\  \n" +
+            "      \\ \\  \\ \\ \\  \\ \\  \\ \\  \\_|\\ \\       \\ \\  \\|\\  \\ \\  \\\\\\  \\   \\ \\  \\ \n" +
+            "       \\ \\__\\ \\ \\__\\ \\__\\ \\_______\\       \\ \\_______\\ \\_______\\   \\ \\__\\\n" +
+            "        \\|__|  \\|__|\\|__|\\|_______|        \\|_______|\\|_______|    \\|__|";
+
+    private static final String PREFIX = "Botty: ";
+    private static final String INDENTATION = "       ";
+    private final Scanner inputScanner;
+
+    public UI() {
+        inputScanner = new Scanner(System.in);
+    }
+    public String getUserInput() {
+        return inputScanner.nextLine();
+    }
+    public void displayIntroduction() {
+        System.out.println(LOGO);
+        System.out.println();
+
+        reply("Hello, I am Botty the Bot, how may I be of service today?");
+    }
+
+    public void reply(String content) {
+        String[] strings = content.split("\n");
+        System.out.println(PREFIX + strings[0]);
+        for (int i = 1; i < strings.length; i++) {
+            System.out.println(INDENTATION + strings[i]);
+        }
+    }
+}
