--- src/main/java/commands/ListCommand.java
@@ -1,12 +1,0 @@
-package commands;
-
-import exceptions.TaskListEmptyException;
-
-import tasks.TaskManager;
-
-public class ListCommand extends Command {
-    @Override
-    public String execute(TaskManager taskManager, ParsedInput parsedInput) throws TaskListEmptyException {
-        return "Here you go!\n" + taskManager.list();
-    }
-}
