--- src/main/java/exceptions/ArgumentNotFoundException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class ArgumentNotFoundException extends BottyException {
-    public ArgumentNotFoundException(String argument) {
-        super("The following argument is not found: " + argument);
-    }
-}
