--- src/main/java/exceptions/BottyException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class BottyException extends Exception {
-    public BottyException(String message) {
-        super(message);
-    }
-}
