--- src/main/java/exceptions/CorruptedTaskStringException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class CorruptedTaskStringException extends BottyException {
-    public CorruptedTaskStringException() {
-        super("Task string is corrupted! Failed to retrieve task");
-    }
-}
