--- src/main/java/exceptions/EmptyArgumentException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class EmptyArgumentException extends BottyException {
-    public EmptyArgumentException(String flag) {
-        super("The following argument is empty: " + flag);
-    }
-}
