--- src/main/java/exceptions/EmptyCommandException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class EmptyCommandException extends BottyException {
-    public EmptyCommandException() {
-        super("commands.Command cannot be empty!");
-    }
-}
