--- src/main/java/exceptions/IncorrectDateFormatException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class IncorrectDateFormatException extends BottyException {
-    public IncorrectDateFormatException(String message) {
-        super(message);
-    }
-}
