--- src/main/java/exceptions/TaskListEmptyException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class TaskListEmptyException extends BottyException {
-    public TaskListEmptyException() {
-        super("Your list is empty! Add a task with the todo, deadline or event command.");
-    }
-}
