--- src/main/java/exceptions/UnknownCommandException.java
@@ -1,7 +1,0 @@
-package exceptions;
-
-public class UnknownCommandException extends BottyException {
-    public UnknownCommandException(String command) {
-        super("I'm sorry, that is not a command I am familiar with. (" + command + ")");
-    }
-}
